catch errorcode 16 in auth response interceptor
