+++ console/src/app/services/grpc.service.ts
@@ -1,6 +1,7 @@
 import { PlatformLocation } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { AuthConfig } from 'angular-oauth2-oidc';
 
 import { AdminServicePromiseClient } from '../proto/generated/admin_grpc_web_pb';
@@ -24,6 +25,8 @@ export class GrpcService {
         private platformLocation: PlatformLocation,
         private authenticationService: AuthenticationService,
         private storageService: StorageService,
+        private dialog: MatDialog,
+        // private toast: ToastService,
     ) { }
 
     public async loadAppEnvironment(): Promise<any> {
@@ -31,9 +34,9 @@ export class GrpcService {
             .toPromise().then((data: any) => {
                 if (data && data.authServiceUrl && data.mgmtServiceUrl && data.issuer) {
                     const interceptors = {
-                        'unaryInterceptors': [
-                            new AuthInterceptor(this.authenticationService, this.storageService),
+                        unaryInterceptors: [
                             new OrgInterceptor(this.storageService),
+                            new AuthInterceptor(this.authenticationService, this.storageService, this.dialog),
                         ],
                     };
