+++ console/src/app/services/interceptors/auth.interceptor.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Request, UnaryInterceptor, UnaryResponse } from 'grpc-web';
-import { filter, first } from 'rxjs/operators';
+import { filter, first, take } from 'rxjs/operators';
+import { WarnDialogComponent } from 'src/app/modules/warn-dialog/warn-dialog.component';
 
 import { AuthenticationService } from '../authentication.service';
 import { StorageService } from '../storage.service';
@@ -10,8 +12,15 @@ const authorizationKey = 'Authorization';
 const bearerPrefix = 'Bearer';
 const accessTokenStorageKey = 'access_token';
 @Injectable({ providedIn: 'root' })
+/**
+ * Set the authentication token
+ */
 export class AuthInterceptor<TReq = unknown, TResp = unknown> implements UnaryInterceptor<TReq, TResp> {
-    constructor(private authenticationService: AuthenticationService, private storageService: StorageService) { }
+    constructor(
+        private authenticationService: AuthenticationService,
+        private storageService: StorageService,
+        private dialog: MatDialog,
+    ) { }
 
     public async intercept(request: Request<TReq, TResp>, invoker: any): Promise<UnaryResponse<TReq, TResp>> {
         await this.authenticationService.authenticationChanged.pipe(
@@ -24,17 +33,25 @@ export class AuthInterceptor<TReq = unknown, TResp = unknown> implements UnaryIn
         metadata[authorizationKey] = `${bearerPrefix} ${accessToken}`;
 
         return invoker(request).then((response: any) => {
-            // const message = response.getResponseMessage();
-            const respMetadata = response.getMetadata();
-
-            // TODO: intercept unauthenticated an authenticate
-
-            // const status = respMetadata['grpc-status'];
-            // console.log(respMetadata, status);
-            // if (status?.code === 16) {
-            //     this.authenticationService.authenticate();
-            // }
             return response;
+        }).catch((error: any) => {
+            console.error('error: ', error);
+            if (error.code === 16) {
+                const dialogRef = this.dialog.open(WarnDialogComponent, {
+                    data: {
+                        confirmKey: 'ACTIONS.LOGIN',
+                        titleKey: 'ERRORS.TOKENINVALID.TITLE',
+                        descriptionKey: 'ERRORS.TOKENINVALID.DESCRIPTION',
+                    },
+                    width: '400px',
+                });
+
+                dialogRef.afterClosed().pipe(take(1)).subscribe(resp => {
+                    if (resp) {
+                        this.authenticationService.authenticate(undefined, true, true);
+                    }
+                });
+            }
         });
     }
 }
